SNOW-1665707: Move general.py into src/…/modin/plugin/extensions/
