+++ src/snowflake/snowpark/modin/pandas/__init__.py
@@ -95,32 +95,6 @@ from snowflake.snowpark.modin.pandas.api.extensions import (
     register_dataframe_accessor,
     register_series_accessor,
 )
-from snowflake.snowpark.modin.pandas.general import (
-    bdate_range,
-    concat,
-    crosstab,
-    cut,
-    date_range,
-    get_dummies,
-    isna,
-    isnull,
-    lreshape,
-    melt,
-    merge,
-    merge_asof,
-    merge_ordered,
-    notna,
-    notnull,
-    pivot,
-    pivot_table,
-    qcut,
-    to_datetime,
-    to_numeric,
-    to_timedelta,
-    unique,
-    value_counts,
-    wide_to_long,
-)
 from snowflake.snowpark.modin.pandas.io import (  # read_json is provided by overrides module
     ExcelFile,
     HDFStore,
@@ -150,6 +124,32 @@ from snowflake.snowpark.modin.plugin._internal.session import SnowpandasSessionH
 from snowflake.snowpark.modin.plugin._internal.telemetry import (
     TELEMETRY_PRIVATE_METHODS,
     try_add_telemetry_to_attribute,
+)
+from snowflake.snowpark.modin.plugin.extensions.general import (
+    bdate_range,
+    concat,
+    crosstab,
+    cut,
+    date_range,
+    get_dummies,
+    isna,
+    isnull,
+    lreshape,
+    melt,
+    merge,
+    merge_asof,
+    merge_ordered,
+    notna,
+    notnull,
+    pivot,
+    pivot_table,
+    qcut,
+    to_datetime,
+    to_numeric,
+    to_timedelta,
+    unique,
+    value_counts,
+    wide_to_long,
 )
 from snowflake.snowpark.modin.plugin.utils.frontend_constants import _ATTRS_NO_LOOKUP
